Improve logic for ObjectPools and Blocks

+++ src/execution_plan/record.h
@@ -30,7 +30,7 @@ typedef struct {
 } Entry;
 
 typedef struct {
-	size_t id;          // Record internal ID.
+	uint id;            // Record internal ID.
 	void *owner;        // Owner of record.
 	rax *mapping;       // Mapping between alias to record entry.
 	Entry entries[];    // Array of entries.
